Separate config extractor files a bit
