+++ src/runconf_ui/interfaces/actions/action_interfaces.py
@@ -1,4 +1,4 @@
-from runconf_ui.interfaces.controller.config_wrapper import ConfigurationWrapper
+from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
 from runconf_ui.exceptions import CiderBadActionException
 
 from abc import ABC, abstractmethod
@@ -18,8 +18,8 @@ class ActionInterface(ABC):
     Generic interface defining an abstract action
     """
 
-    def __init__(self, configuration: ConfigurationWrapper):
-        self._configuration = configuration
+    def __init__(self, configuration: DaqConfigurationWrapper):
+        self._daq_configuration = configuration
 
     @abstractmethod
     def action(self, *args, **kwargs):
@@ -34,4 +34,4 @@ class ActionInterface(ABC):
             raise CiderBadActionException(traceback.format_exc())
 
     def __str__(self):
-        return f"{self.__class__.__name__} using {self._configuration.file_name}"
+        return f"{self.__class__.__name__} using {self._daq_configuration.file_name}"
