+++ src/runconf_ui/interfaces/workflows/get_objects_in_session.py
@@ -10,11 +10,11 @@ class GetSegmentAppsListAction(ActionInterface):
         apps_list_flat = []
 
         # Loop over all segments
-        for ss in ca.GetAttributeAction(self._configuration)(segment, "segments"):
+        for ss in ca.GetAttributeAction(self._daq_configuration)(segment, "segments"):
             apps_list_flat += self.action(ss)
 
         # loop over all attributes
-        for aa in ca.GetAttributeAction(self._configuration)(segment, "applications"):
+        for aa in ca.GetAttributeAction(self._daq_configuration)(segment, "applications"):
             apps_list_flat.append(aa)
 
         return apps_list_flat
@@ -27,21 +27,21 @@ class GetObjectsInSessionAction(ActionInterface):
 
     def action(self, session_dal, applied_class: str, specific_objects=None):
         # Session contains a single segement
-        segment = ca.GetAttributeAction(self._configuration)(session_dal, "segment")
+        segment = ca.GetAttributeAction(self._daq_configuration)(session_dal, "segment")
         # Get all apps+segments in the segment
-        full_app_list = GetSegmentAppsListAction(self._configuration)(segment)
+        full_app_list = GetSegmentAppsListAction(self._daq_configuration)(segment)
 
         apps = []
         for app in full_app_list:
             # Check if we have some subset of object
             if (
                 specific_objects is not None
-                and ca.GetAttributeAction(self._configuration)(app, "id")
+                and ca.GetAttributeAction(self._daq_configuration)(app, "id")
                 not in specific_objects
             ):
                 continue
             # Check if the object is of the right class
-            if ca.GetClassNameAction(self._configuration)(app) != applied_class:
+            if ca.GetClassNameAction(self._daq_configuration)(app) != applied_class:
                 continue
             apps.append(app)
         return apps
