+++ src/runconf_ui/utils/daq_conf_tools/daq_conf_path_reader.py
@@ -1,5 +1,5 @@
 import runconf_ui.interfaces.actions.actions as ca
-from runconf_ui.interfaces.controller.config_wrapper import ConfigurationWrapper
+from runconf_ui.interfaces.controller.daq_conf_wrapper import DaqConfigurationWrapper
 
 from pathlib import Path
 from typing import List
@@ -36,7 +36,7 @@ class DaqConfPathReader:
 
         return file_path
 
-    def __check_default_session(self, session, config_file: ConfigurationWrapper) -> bool:
+    def __check_default_session(self, session, config_file: DaqConfigurationWrapper) -> bool:
         if self._default_session is None:
             return True
 
@@ -45,7 +45,7 @@ class DaqConfPathReader:
     def _get_number_of_sessions(self, config_file_path: str) -> int:
         """Returns the number of sessions in the given configuration file."""
         try:
-            config_file = ConfigurationWrapper(config_file_path)
+            config_file = DaqConfigurationWrapper(config_file_path)
 
             return len(
                 [
