+++ src/runconf_ui/widgets/enable_disable_base.py
@@ -1,4 +1,4 @@
-from runconf_ui.interfaces.workflows.extract_system_info import SubsystemStatus
+from runconf_ui.utils.shifter_config_tools.daq_system_readers. import SubsystemStatus
 
 from textual.visual import SupportsVisual
 from textual.widgets import Static, Button
