+++ src/runconf_ui/widgets/multicomponent_panel.py
@@ -1,8 +1,7 @@
 from runconf_ui.widgets.enable_disable_base import EnableDisablePanel
-from runconf_ui.interfaces.workflows.extract_system_info import (
-    SubsystemStatus,
-    DetectorExtractor,
-)
+from runconf_ui.utils.shifter_config_tools.daq_system_readers.detector_extractor import DetectorExtractor
+from runconf_ui.utils.subsystem_status import SubsystemStatus
+
 from runconf_ui.utils.daq_conf_tools.daq_conf_tree import ComponentLevelTree
 from runconf_ui.interfaces.controller.application_controller import (
     ShifterInterfaceState,
@@ -50,7 +49,7 @@ class MultiComponentEnableDisablePanel(EnableDisablePanel):
         self._disabled_items = []
 
         self._extractor = DetectorExtractor(
-            self._application_controller.dummy_oks_configuration,
+            self._application_controller.buffer_daq_config,
             self._application_controller.session_name,
             object_list,
         )
@@ -58,13 +57,13 @@ class MultiComponentEnableDisablePanel(EnableDisablePanel):
     def generate_button_list(self) -> Dict | None:
         if (
             self._application_controller.session_name is None
-            or self._application_controller.dummy_oks_configuration is None
+            or self._application_controller.buffer_daq_config is None
         ):
             return {}
 
         # Set up information extractor
         self._extractor.set_config_session(
-            self._application_controller.dummy_oks_configuration,
+            self._application_controller.buffer_daq_config,
             self._application_controller.session_name,
         )
